requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.2.0
requests==2.31.0
python-dotenv==1.0.0
boto3==1.34.34
sagemaker==2.199.0
joblib==1.3.2
